Use Numpy's any to avoid error on None
Using Python3.3, I got an error when doing

pcolormesh(C, norm=SymLogNorm(1.0))

It complained that 'None' was not iterable for:

if any(xticklabels):
    ...

I did not experience it on Python2.7, so perhaps they
changed the implementation of 'any'. In any case, using
Numpy's 'any' function fixes the issue because it correctly
returns False if the object is None.

Signed-off-by: Jonas Kalderstam <jonas@kalderstam.se>